small refactoring update ULFileSystem implement ULBuffer ported.cs add Buffer.MemoryCopy JSString.UTF16Data is now char ULRect : IEquatable<ULRect>, readonly bool IsEquals ULIntRect : IEquatable<ULIntRect> Use const DllImport names make ULGPUDriver IEquatable, short down setters, implement GetHashCode, make getters readonly make ULLogger IEquatable, short down setters, implement GetHashCode, make getters readonly make ULSurfaceDefinition IEquatable, implement GetHashCode, make getters readonly specify enum size source make ULClipboard IEquatable, short down setters, readonly getters, GetHashCode Use _ULConfig* in c api Update ULConfig c api to latest changes, specify missing default values Mark methods readonly replace ULStringDisposableStackToNativeMarshaller, ULStringToNativeMarshaller, ULStringGeneratedDllImportMarshaler with ULString, ULString.ToNative, ULString.ToManaged_ make ULBitmap IEquatable, implement Equals + ReferenceEquals update benchmarj=ks

+++ src/UltralightNet/Enums/ULKeyEventModifiers.cs
@@ -4,7 +4,7 @@ namespace UltralightNet;
 
 /// <summary>An enumeration of the different keyboard modifiers.</summary>
 [Flags]
-public enum ULKeyEventModifiers : uint
+public enum ULKeyEventModifiers : byte // KeyEvent.h - uint8_t, CAPI_KeyEvent.h - unsigned int
 {
 	/// <summary>Whether or not an ALT key is down</summary>
 	AltKey = 1 << 0,
